feat(gate1): delete watchdog + recipient_liveness (cross-agent liveness push)
Gate 1 of the Hybrid migration — the least-entangled push deletion (codex-narrowed).
-1688/+27 LOC. Tree green; conformance green.

DELETE: watchdog.go (cross-agent liveness sweep / cooperative poke daemon),
recipient_liveness.go (evaluateRecipientLiveness cluster), + their tests.
CO-TRIM (intended cross-agent-liveness removal, not regressions):
- main.go: drop `watchdog` verb dispatch + help.
- check.go: drop cooperative-waking / liveness-sweep block + orphaned context import
  (inbox read/quarantine/archive/reply-obligation recording untouched).
- gate.go: drop recipient-liveness evaluation + LivenessOK/Message fields + the
  liveness blocking/warning block. Own-obligation gating (unread/malformed/pending/
  corrupt-ledger/missing-reg/stale-reg) fully retained. Gate no longer blocks on
  cross-agent recipient liveness (intended per the decision).
- doctor.go: drop checkRecipientLiveness; framework + other checks intact.
- reachability.go: relocate registrationHasReachableWake into it (out of deleted
  recipient_liveness.go) so poller.go stays untouched.

DEFERRED (correct, per entanglement): reachability.go deletion → Gate 6 (its
reproveAndRebindOwnWake is called from run.go:685, off-limits this gate).

Compile gate (codex): no cmdWatchdog/checkRecipientLiveness/evaluateRecipientLiveness/
runCooperativeWaking runtime refs remain.
